Fix s:GetFileWinnr to use file path
This was causing issues when combining tagbar with coc & ccls using
references, where coc creates and destroys preview windows. Somewhere
tagbar remembers the bufnr of the preview windows instead of the
file-path and later throws errors when trying to browse using the
tagbar.

This should fix it; however, there can be unintentional consequences.